Store private key encrypyted
After generating a keypair, the private key is stored encrypted using
a DES key, generated from a password. When the private key is required
(while encrypting or decrypting messages) the private key will be decrypted.

The password used to encrypt the private key is never stored. Also, the
private key will only be decrypted on the fly. Its decrypted form will
never be stored or remembered.