Added specify recipients activity, activated when sending to multiple recipients
When sending to multiple recipients, and for some we have their public key,
and for some we don't, then show the specify recipients activity. This
activity shows a list to who the message will be sent encrypted, and to who
it will be sent in plain text.

In the list, the user is able to uncheck recipients, to make sure messages
are not sent in plaintext by accident. After submitting the form, the message
will be sent to only the recipients that are checked.