Added encrypted message sms-count and return if a message is sent or not
Refactored MultipartDataMessage to an object, with methods to get the
message part count and a send method. The send method returns a true if
a message is sent and false if there was a problem. This boolean is used
in serveral other places to determine the result of sending a
datamessage.